Move helper functions for HDISO sources to pyanaconda.ui.lib
These functions are used only in GUI and TUI.
