+++ pyanaconda/payload/image.py
@@ -25,13 +25,10 @@ import tempfile
 import blivet.util
 import blivet.arch
 
-from blivet.size import Size
-
 from pyanaconda import isys
 from pyanaconda.modules.common.constants.objects import DEVICE_TREE
 from pyanaconda.modules.common.constants.services import STORAGE
 from pyanaconda.modules.common.errors.storage import MountFilesystemError
-from pyanaconda.modules.common.structures.storage import DeviceData, DeviceFormatData
 from pyanaconda.payload import utils as payload_utils
 from pyanaconda.modules.payloads.payload.dnf.tree_info import TreeInfoMetadata, \
     TreeInfoMetadataError
@@ -174,52 +171,3 @@ def find_optical_install_media():
         return dev
 
     return None
-
-
-def find_potential_hdiso_sources():
-    """Find potential HDISO sources.
-
-    Return a generator yielding Device instances that may have HDISO install
-    media somewhere. Candidate devices are simply any that we can mount.
-
-    :return: a list of device names
-    """
-    device_tree = STORAGE.get_proxy(DEVICE_TREE)
-    return device_tree.FindMountablePartitions()
-
-
-def get_hdiso_source_info(device_tree, device_name):
-    """Get info about a potential HDISO source.
-
-    :param device_tree: a proxy of a device tree
-    :param device_name: a device name
-    :return: a dictionary with a device info
-    """
-    device_data = DeviceData.from_structure(
-        device_tree.GetDeviceData(device_name)
-    )
-
-    format_data = DeviceFormatData.from_structure(
-        device_tree.GetFormatData(device_name)
-    )
-
-    disk_data = DeviceData.from_structure(
-        device_tree.GetDeviceData(device_data.parents[0])
-    )
-
-    return {
-        "model": disk_data.attrs.get("model", "").replace("_", " "),
-        "path": device_data.path,
-        "size": Size(device_data.size),
-        "format": format_data.description,
-        "label": format_data.attrs.get("label") or format_data.attrs.get("uuid") or ""
-    }
-
-
-def get_hdiso_source_description(device_info):
-    """Get a description of a potential HDISO source.
-
-    :param device_info: a dictionary with a device info
-    :return: a string with a device description
-    """
-    return "{model} {path} ({size}) {format} {label}".format(**device_info)
