+++ pyanaconda/ui/lib/payload.py
@@ -15,10 +15,13 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
+from blivet.size import Size
 from dasbus.client.proxy import get_object_path
 from pyanaconda.anaconda_loggers import get_module_logger
 from pyanaconda.core.constants import PAYLOAD_TYPE_DNF
-from pyanaconda.modules.common.constants.services import PAYLOADS
+from pyanaconda.modules.common.constants.objects import DEVICE_TREE
+from pyanaconda.modules.common.constants.services import PAYLOADS, STORAGE
+from pyanaconda.modules.common.structures.storage import DeviceData, DeviceFormatData
 from pyanaconda.modules.common.task import sync_run_task
 
 log = get_module_logger(__name__)
@@ -136,3 +139,52 @@ def tear_down_sources(payload_proxy):
     task_path = payload_proxy.TearDownSourcesWithTask()
     task_proxy = PAYLOADS.get_proxy(task_path)
     sync_run_task(task_proxy)
+
+
+def find_potential_hdiso_sources():
+    """Find potential HDISO sources.
+
+    Return a generator yielding Device instances that may have HDISO install
+    media somewhere. Candidate devices are simply any that we can mount.
+
+    :return: a list of device names
+    """
+    device_tree = STORAGE.get_proxy(DEVICE_TREE)
+    return device_tree.FindMountablePartitions()
+
+
+def get_hdiso_source_info(device_tree, device_name):
+    """Get info about a potential HDISO source.
+
+    :param device_tree: a proxy of a device tree
+    :param device_name: a device name
+    :return: a dictionary with a device info
+    """
+    device_data = DeviceData.from_structure(
+        device_tree.GetDeviceData(device_name)
+    )
+
+    format_data = DeviceFormatData.from_structure(
+        device_tree.GetFormatData(device_name)
+    )
+
+    disk_data = DeviceData.from_structure(
+        device_tree.GetDeviceData(device_data.parents[0])
+    )
+
+    return {
+        "model": disk_data.attrs.get("model", "").replace("_", " "),
+        "path": device_data.path,
+        "size": Size(device_data.size),
+        "format": format_data.description,
+        "label": format_data.attrs.get("label") or format_data.attrs.get("uuid") or ""
+    }
+
+
+def get_hdiso_source_description(device_info):
+    """Get a description of a potential HDISO source.
+
+    :param device_info: a dictionary with a device info
+    :return: a string with a device description
+    """
+    return "{model} {path} ({size}) {format} {label}".format(**device_info)
